reorganise combinations function into utils
