+++ src/chemnlp/data_val/config.py
@@ -1,4 +1,3 @@
-import itertools
 from typing import Any, Dict, List, Optional, Union
 
 from pydantic import BaseModel, validator
@@ -94,13 +93,3 @@ class GridSearch(BaseModel):
     prompt_tuning: Optional[DictofLists] = {}
     trainer: Optional[DictofLists] = {}
     wandb: Optional[DictofLists] = {}
-
-
-def _get_all_combinations(d: Dict):
-    """Generate all possible hyperparameter combinations"""
-    keys, values = d.keys(), d.values()
-    values_choices = (
-        _get_all_combinations(v) if isinstance(v, dict) else v for v in values
-    )
-    for comb in itertools.product(*values_choices):
-        yield dict(zip(keys, comb))
