+++ src/chemnlp/utils.py
@@ -1,5 +1,6 @@
 from pathlib import Path
-from typing import Union
+import itertools
+from typing import Dict, Union
 
 import yaml
 
@@ -10,3 +11,12 @@ def load_config(path: Union[str, Path]):
             return yaml.safe_load(stream)
         except yaml.YAMLError as exc:
             print(exc)
+
+def _get_all_combinations(d: Dict):
+    """Generate all possible hyperparameter combinations"""
+    keys, values = d.keys(), d.values()
+    values_choices = (
+        _get_all_combinations(v) if isinstance(v, dict) else v for v in values
+    )
+    for comb in itertools.product(*values_choices):
+        yield dict(zip(keys, comb))
